requirements.txt
aiogram~=3.13.0
asyncio~=3.4.3
psycopg2~=2.9.9
aiohttp~=3.9.5
python-dotenv